// generated/src/app/api-models/cap-pool-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapPoolSelect {
  capModelId: string;
  capEntityCode: string;
  capEntityId: string;
  seqCapPoolId: number;
  seqCapPoolSel: number;
  columnName: string;
  columnOccurrence: number;
  columnType: string;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  capPoolId: string;
}